refactor(colors): Rename colors to mapColorUtil
