+++ src/components/TreesMap/mapColorUtil.ts
